requirements.txt
pandas==2.0.3
sqlalchemy==2.0.19
pyyaml==6.0.1
jsonschema==4.19.0
loguru==0.7.0
pytest==7.4.0
click==8.1.6
